backend/scripts/python-dashboard-files/requirements.txt
requests==2.31.0
python-dotenv==1.0.0
customtkinter==5.2.0 